Rename BabySqueel::Association::AliasingError to BabySqueel::AssociationAliasingError
